+++ src/modules/missions/controllers/mission.controller.ts
@@ -1,14 +1,29 @@
-import { Request, Response, NextFunction } from "express";
-import { StatusCodes } from "http-status-codes";
-import { addMissionRequest, IAddMissionRequest } from "../dtos/mission.dto.js";
-import { addStoreMission } from "../services/mission.service.js";
+import { IAddMissionRequest, IChallengeRequest, IChallengeResponse, IMissionResponse } from "../dtos/mission.dto.js";
+import { addStoreMission, challengeStoreMission } from "../services/mission.service.js";
+import { Body, Controller, Post, Route, Tags } from "tsoa";
+import { ApiResponse, success } from "../../../common/responses/response.js";
 
 // 가게에 미션 추가하기
-export const handleAddMission = async (req: Request, res: Response, next: NextFunction) => {
-  console.log("가게 미션 추가를 요청했습니다.");
-  console.log("body: ", req.body);
+@Route("missions")
+@Tags("Missions")
+export class MissionController extends Controller {
+  // 가게에 미션 추가하기
+  @Post("")
+  public async handleAddMission(@Body() body: IAddMissionRequest): Promise<ApiResponse<IMissionResponse>> {
+    console.log("가게 미션 추가를 요청했습니다.");
+    console.log("body: ", body);
 
-  const mission = await addStoreMission(addMissionRequest(req.body as IAddMissionRequest));
+    const mission = await addStoreMission(body);
+    return success(mission);
+  }
 
-  res.status(StatusCodes.OK).json({ result: mission });
-};
+  // 미션 도전하기
+  @Post("challenge")
+  public async handleChallengeMission(@Body() body: IChallengeRequest): Promise<ApiResponse<IChallengeResponse>> {
+    console.log("미션 도전을 요청했습니다.");
+    console.log("body: ", body);
+
+    const challenge = await challengeStoreMission(body);
+    return success(challenge);
+  }
+}
