+++ src/modules/missions/dtos/challenge.dto.ts
@@ -4,35 +4,6 @@ export enum MissionStatus {
   COMPLETED = "진행완료",
 }
 
-// 미션 도전하기
-export interface IChallengeRequest {
-  memberId: number;
-  missionId: number;
-}
-
-export const challengeRequest = (body: IChallengeRequest) => {
-  return {
-    memberId: body.memberId,
-    missionId: body.missionId,
-    status: MissionStatus.IN_PROGRESS,
-  };
-};
-
-// 미션 조회하기
-export interface IChallengeResponse {
-  memberId: number;
-  missionId: number;
-  status: MissionStatus;
-}
-
-export const challengeResponse = (body: IChallengeResponse) => {
-  return {
-    memberId: body.memberId,
-    missionId: body.missionId,
-    status: body.status,
-  };
-};
-
 // 미션 목록 조회
 export interface ChallengeItem {
   challengeId: number;
