+++ src/modules/missions/repositories/mission.repository.ts
@@ -1,4 +1,5 @@
 import { prisma } from "../../../db.config.js";
+import { MissionStatus } from "../dtos/challenge.dto.js";
 
 // 가게에 미션 추가하기
 export const addMission = async (data: any) => {
@@ -19,4 +20,32 @@ export const getMission = async (missionId: number) => {
   const mission = await prisma.mission.findUnique({ where: { id: missionId } });
 
   return mission;
+};
+
+// 미션 도전하기
+export const challengeMission = async (data: any) => {
+  // 이미 도전 중인지 확인
+  const confirm = await prisma.member_mission.findFirst({ where: { member_id: data.memberId, mission_id: data.missionId } });
+
+  if (confirm) {
+    return null;
+  }
+
+  // 미션 도전하기
+  const challenge = await prisma.member_mission.create({
+    data: {
+      member_id: data.memberId,
+      mission_id: data.missionId,
+      status: MissionStatus.IN_PROGRESS,
+    },
+  });
+
+  return challenge.id;
+};
+
+// 도전 중인 미션 개별조회
+export const getChallenge = async (challengeId: number) => {
+  const challenge = await prisma.member_mission.findUnique({ where: { id: challengeId } });
+
+  return challenge;
 };
